// src/app/modules/staff/model/staff-member-filter-view.ts
/**
 * StaffWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { PagingParams } from 'app/models/paging-params';
import { SortOrder } from 'app/models/sort-order';

/**StaffMamberFilterView */
export interface StaffMemberFilterView extends PagingParams {
    /**Gets or Sets SortBy */
    sortBy: string;
    /**Gets or Sets StatusIds */
    statusIds?: string[];
    /**Gets or Sets StaffNames */
    staffNames?: string[];
    /**Gets or Sets StaffCodes */
    staffCodes?: string[];
    /**Gets or Sets Departments */
    departments?: string[];
    /**Gets or Sets StaffTypeIds */
    staffTypeIds?: string[];
    /**Gets or Sets SortOrder */
    sortOrder: SortOrder;
}