// src/app/modules/staff/staff-information-system/model/staff-status-view.ts
/**
 * Teacher API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * StaffStatusView
 */
export interface StaffStatusView { 
    /**
     * Gets or sets column StaffId
     */
    staffId?: string;
    /**
     * Gets or sets column StatusId
     */
    statusId?: string;
}